Trim dead profile storage wrappers
